ensure that string encoded float lists are parsed with atof, not atoi
git-svn-id: https://svn.osgeo.org/gdal/trunk@23414 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrfeature.cpp
@@ -1970,7 +1970,7 @@ void OGRFeature::SetField( int iField, const char * pszValue )
             std::vector<double> adfValues;
 
             for( i=0; i < nCount; i++ )
-                adfValues.push_back( atoi(papszValueList[i+1]) );
+                adfValues.push_back( atof(papszValueList[i+1]) );
             SetField( iField, nCount, &(adfValues[0]) );
         }
